Migrate details
